Move 18Cz to new structure
